requirements.txt
apscheduler
astral
cairosvg
filelock
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gpiozero
lgpio
Pillow
paho-mqtt
python-dateutil
PyYAML
requests
RPi.GPIO
spidev
